Added the Arduino Shield I designed
This commit adds the files for the design of the Arduino shield that connects all the different parts of the hardware side of this project. It is designed to get power from a standard USB 5V charger, and to use this to run an Arduino Uno, an HC-06 bluetooth module, and a relay that toggles power to the output USB port, that is designed to charge slowly a smartphone, that would control the contraption via Bluetooth, using the app.